+++ include/boost/beast/core/detail/type_traits.hpp
@@ -11,6 +11,7 @@
 #define BOOST_BEAST_DETAIL_TYPE_TRAITS_HPP
 
 #include <boost/beast/core/error.hpp>
+#include <boost/beast/core/detail/is_invocable.hpp>
 #include <boost/asio/buffer.hpp>
 #include <boost/mp11/function.hpp>
 #include <boost/type_traits/make_void.hpp>
@@ -99,42 +100,6 @@ using aligned_union_t =
 template<class T>
 void
 accept_rv(T){}
-
-//------------------------------------------------------------------------------
-
-template<class R, class C, class ...A>
-auto
-is_invocable_test(C&& c, int, A&& ...a)
-    -> decltype(std::is_convertible<
-        decltype(c(std::forward<A>(a)...)), R>::value ||
-            std::is_same<R, void>::value,
-                std::true_type());
-
-template<class R, class C, class ...A>
-std::false_type
-is_invocable_test(C&& c, long, A&& ...a);
-
-/** Metafunction returns `true` if F callable as R(A...)
-
-    Example:
-
-    @code
-        is_invocable<T, void(std::string)>
-    @endcode
-*/
-/** @{ */
-template<class C, class F>
-struct is_invocable : std::false_type
-{
-};
-
-template<class C, class R, class ...A>
-struct is_invocable<C, R(A...)>
-    : decltype(is_invocable_test<R>(
-        std::declval<C>(), 1, std::declval<A>()...))
-{
-};
-/** @} */
 
 //------------------------------------------------------------------------------
