+++ include/boost/beast/core/type_traits.hpp
@@ -12,6 +12,7 @@
 
 #include <boost/beast/core/detail/config.hpp>
 #include <boost/beast/core/file_base.hpp>
+#include <boost/beast/core/detail/is_invocable.hpp>
 #include <boost/beast/core/detail/type_traits.hpp>
 #include <boost/asio/buffer.hpp>
 #include <type_traits>
