+++ test/beast/core/Jamfile
@@ -11,6 +11,7 @@ local SOURCES =
     _detail_base64.cpp
     _detail_buffer.cpp
     _detail_clamp.cpp
+    _detail_is_invocable.cpp
     _detail_read.cpp
     _detail_sha1.cpp
     _detail_tuple.cpp
