+++ test/beast/core/type_traits.cpp
@@ -19,50 +19,6 @@
 namespace boost {
 namespace beast {
 
-namespace detail {
-
-namespace {
-
-//
-// is_invocable
-//
-
-struct is_invocable_udt1
-{
-    void operator()(int) const;
-};
-
-struct is_invocable_udt2
-{
-    int operator()(int) const;
-};
-
-struct is_invocable_udt3
-{
-    int operator()(int);
-};
-
-struct is_invocable_udt4
-{
-    void operator()(std::unique_ptr<int>);
-};
-
-#ifndef __INTELLISENSE__
-// VFALCO Fails to compile with Intellisense
-BOOST_STATIC_ASSERT(is_invocable<is_invocable_udt1, void(int)>::value);
-BOOST_STATIC_ASSERT(is_invocable<is_invocable_udt2, int(int)>::value);
-BOOST_STATIC_ASSERT(is_invocable<is_invocable_udt3, int(int)>::value);
-BOOST_STATIC_ASSERT(! is_invocable<is_invocable_udt1, void(void)>::value);
-BOOST_STATIC_ASSERT(! is_invocable<is_invocable_udt2, int(void)>::value);
-BOOST_STATIC_ASSERT(! is_invocable<is_invocable_udt2, void(void)>::value);
-BOOST_STATIC_ASSERT(! is_invocable<is_invocable_udt3 const, int(int)>::value);
-BOOST_STATIC_ASSERT(is_invocable<is_invocable_udt4, void(std::unique_ptr<int>)>::value);
-#endif
-
-} // (anonymous)
-
-} // detail
-
 //
 // handler concepts
 //
@@ -71,7 +27,7 @@ namespace {
 
 struct H
 {
-    void operator()(int);
+    void operator()(int){}
 };
 
 } // anonymous
